test(memory): extract contract test suite, add Postgres integration tests
